FGD-37/Revamped Monologue Map

+++ Assets/Scripts/Player/MonologueOnAwake.cs
@@ -4,6 +4,6 @@ public class MonologueOnAwake : MonoBehaviour {
     public MonologueKey monologueKey;
 
     private void Start() {
-        GetComponent<MonologueUI>().StartMonologue((int)monologueKey);
+        GetComponent<MonologueUI>().StartMonologue(monologueKey);
     }
 }
